+++ lib/netzke/basepack/item_persistence.rb
@@ -1,7 +1,7 @@
 module Netzke
   module Basepack
     # When mixed into a component with resizable layout (e.g. border layout), this module enables persistence for regions size and collapsed/expanded state.
-    module ItemsPersistence
+    module ItemPersistence
       extend ActiveSupport::Concern
 
       included do