+++ test/basepack_test_app/app/components/panel_with_persistent_regions.rb
@@ -1,5 +1,5 @@
 class PanelWithPersistentRegions < Netzke::Base
-  include Netzke::Basepack::ItemsPersistence
+  include Netzke::Basepack::ItemPersistence
 
   js_configure do |c|
     c.layout = :border